requirements.txt
# Core FastAPI and server dependencies - Python 3.12 compatible
fastapi>=0.115.0
uvicorn[standard]>=0.23.2
gunicorn>=21.2.0
websockets>=12.0
python-multipart>=0.0.9

# Database dependencies - Python 3.12 compatible
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Data validation and security - Python 3.12 compatible
pydantic>=2.5.3
bcrypt>=4.1.2
python-dotenv>=1.0.0

# Essential AI/LLM dependencies - Python 3.12 compatible
# Using older langchain version for compatibility with existing imports
langchain==0.1.20
langchain-openai>=0.1.0
openai>=1.0.0
pinecone>=3.0.0

# HTTP and API dependencies - Python 3.12 compatible
requests>=2.31.0
httpx>=0.25.2

# Document processing - Python 3.12 compatible
pdfminer.six>=20231228
PyPDF2>=3.0.0  # PDF text extraction for protocol extraction
pdfplumber>=0.10.0  # Fallback PDF extraction

# Utilities - Python 3.12 compatible
jsonschema>=4.20.0
sendgrid>=6.11.0

# Basic scientific computing (no heavy ML dependencies) - Python 3.12 compatible
numpy>=1.26.0

# Phase 2A: NLP Infrastructure for Semantic Analysis - Basic NLP only for deployment stability
nltk>=3.8.1
spacy>=3.7.0
# Advanced NLP dependencies (commented out for deployment stability)
# transformers>=4.35.0
# torch>=2.1.0
# sentence-transformers>=2.2.2

# Phase 5: Citation Network Foundation dependencies - Python 3.12 compatible
scikit-learn>=1.3.0
aiohttp>=3.8.0
